+++ rastervision_core/rastervision/core/cli.py
@@ -39,7 +39,13 @@ class OptionEatAll(click.Option):
     'predict', short_help='Use a model bundle to predict on new images.')
 @click.argument('model_bundle')
 @click.argument('image_uri')
-@click.argument('output_uri')
+@click.argument('label_uri')
+@click.option(
+    '--vector-label-uri',
+    type=str,
+    help=
+    ('URI to save vectorized labels for semantic segmentation model bundles that support '
+     'it'))
 @click.option(
     '--update-stats',
     '-a',
@@ -51,9 +57,10 @@ class OptionEatAll(click.Option):
     '--channel-order',
     cls=OptionEatAll,
     help='List of indices comprising channel_order. Example: 2 1 0')
-def predict(model_bundle, image_uri, output_uri, update_stats, channel_order):
+def predict(model_bundle, image_uri, label_uri, vector_label_uri, update_stats,
+            channel_order):
     """Make predictions on the images at IMAGE_URI
-    using MODEL_BUNDLE and store the prediction output at OUTPUT_URI.
+    using MODEL_BUNDLE and store the prediction output at LABEL_URI.
     """
     if channel_order is not None:
         channel_order = [
@@ -63,4 +70,5 @@ def predict(model_bundle, image_uri, output_uri, update_stats, channel_order):
     with rv_config.get_tmp_dir() as tmp_dir:
         predictor = Predictor(model_bundle, tmp_dir, update_stats,
                               channel_order)
-        predictor.predict([image_uri], output_uri)
+        predictor.predict(
+            [image_uri], label_uri, vector_label_uri=vector_label_uri)